Delete and Edit + HTML/JSP Polish
I added a few servlets and updated the Helpers to add Delete and Edit functionality for lists. The application is now technically complete, though we should have done some things differently. HTML and JSP have had some CSS added to make it a little more appealing to interact with. All we need is our AWS setup and we're good.